requirements.txt
-i https://pypi.org/simple
alembic==1.9.2; python_version >= '3.7'
boto3==1.28.59; python_version >= '3.7'
botocore==1.31.59; python_version >= '3.7'
click==8.1.3; python_version >= '3.7'
flask==2.2.2; python_version >= '3.7'
flask-cors==3.0.10
flask-login==0.6.2; python_version >= '3.7'
flask-migrate==4.0.2; python_version >= '3.6'
flask-sqlalchemy==3.0.2; python_version >= '3.7'
flask-wtf==1.1.1; python_version >= '3.7'
greenlet==2.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
gunicorn==20.1.0; python_version >= '3.5'
importlib-metadata==6.7.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0; python_version >= '3.7'
python-editor==1.0.4
s3transfer==0.7.0; python_version >= '3.7'
setuptools==68.2.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.46; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
urllib3==1.26.17; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2; python_version >= '3.7'
wtforms==3.0.1; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
requests==2.31.0
rsa==4.9
setuptools==68.0.0
urllib3==1.26.16
